Improve leagibility

+++ src/utility/heap.cc
@@ -7,7 +7,7 @@ extern "C" { void _panic(); }
 __BEGIN_UTIL
 
 // Methods
-void Heap::out_of_memory(unsigned int bytes)
+void Heap::out_of_memory(unsigned long bytes)
 {
     db<Heaps, System>(ERR) << "Heap::alloc(this=" << this << "): out of memory while allocating " << bytes << " bytes!" << endl;
